Pet profile management + corrections

+++ src/components/ProductInfo.js
@@ -1,16 +1,32 @@
-/* eslint-disable */
+/* eslint-disable no-console */
 import { useContext, useEffect, useState } from 'react';
+import { useToasts } from 'react-toast-notifications';
 import qs from 'query-string';
+import { NavLink } from 'react-router-dom';
+import {
+  FacebookShareButton,
+  TwitterShareButton,
+  FacebookIcon,
+  TwitterIcon,
+  FacebookShareCount,
+} from 'react-share';
 import API from '../APIClient';
 import { FoodContext } from '../contexts/FoodContext';
 import { CurrentUserContext } from '../contexts/CurrentUserContext';
+import DetailsRating from './DetailsRating';
+import ModalSignIn from './modalSignIn';
 
 export default function ProductInfo() {
   const { foodDetails, setFoodDetails } = useContext(FoodContext);
   const { id } = qs.parse(window.location.search);
-  const { profile, toggleFoodInFavorites, favoritesIdsList } =
-    useContext(CurrentUserContext);
+  const {
+    profile,
+    toggleFoodInFavorites,
+    favoritesIdsList,
+    isLoggedIn = true,
+  } = useContext(CurrentUserContext);
   const [statsInfos, setStatsInfos] = useState(null);
+  const { addToast } = useToasts();
 
   useEffect(async () => {
     API.get(`/foods/${id}`)
@@ -60,40 +76,59 @@ export default function ProductInfo() {
   const handleClickFavorite = async () => {
     const isFavorite = !!favoritesIdsList[id];
     const foodId = parseInt(id, 10);
-
-    if (isFavorite) {
-      API.delete(`/favorites/${foodId}`)
-        .then(async () => {
-          toggleFoodInFavorites(foodId);
-          setStatsInfos({
-            ...statsInfos,
-            foodId,
-            requestInfo: 'removeFavorite',
-            requestSentAt: new Date(),
-          });
-        })
-        .catch((err) => console.log(err));
+    if (profile) {
+      if (isFavorite) {
+        API.delete(`/favorites/${foodId}`)
+          .then(async () => {
+            toggleFoodInFavorites(foodId);
+            setStatsInfos({
+              ...statsInfos,
+              foodId,
+              requestInfo: 'removeFavorite',
+              requestSentAt: new Date(),
+            });
+          })
+          .catch((err) => console.log(err));
+      } else {
+        API.post(`/favorites`, { foodId })
+          .then(async () => {
+            toggleFoodInFavorites(foodId);
+            setStatsInfos({
+              ...statsInfos,
+              foodId,
+              requestInfo: 'addFavorite',
+              requestSentAt: new Date(),
+            });
+          })
+          .catch((err) => console.log(err));
+      }
     } else {
-      API.post(`/favorites`, { foodId })
-        .then(async () => {
-          toggleFoodInFavorites(foodId);
-          setStatsInfos({
-            ...statsInfos,
-            foodId,
-            requestInfo: 'addFavorite',
-            requestSentAt: new Date(),
-          });
-        })
-        .catch((err) => console.log(err));
+      addToast('Vous devez être connecté pour mettre un aliment en favori !', {
+        appearance: 'error',
+      });
     }
   };
+
+  const handleNotConnected = () => {
+    if (!profile) {
+      addToast(
+        'Vous devez être connecté pour donner votre avis sur un aliment !',
+        {
+          appearance: 'error',
+        }
+      );
+    }
+  };
+
+  const shareUrl = `${process.env.REACT_APP_FRONT_BASE_URL}/product-info-page/?id=${id}`;
+  const title = 'Gamelle : ';
 
   return (
     <>
       {foodDetails && (
         <>
           <div className="flex items-center flex-col justify-center md:p-5">
-            <div className="relative md:flex md:flex-col md:shadow-lg lg:w-7/12 md:w-10/12 md:m-10 bg-white dark:bg-darkpurple">
+            <div className="relative md:flex md:flex-col md:shadow-lg lg:w-10/12 md:w-10/12 md:m-10 bg-white">
               <div className="absolute right-0 mr-5 mt-3">
                 <button
                   type="button"
@@ -118,8 +153,55 @@ export default function ProductInfo() {
                   <div className="font-bold text-4xl">{foodDetails.brand}</div>
                   <div className="text-base">{foodDetails.name}</div>
                 </div>
-              </div>
+                <div className="flex flex-col w-full right-0 items-end">
+                  <div className="w-2/3 flex justify-center flex-col items-center">
+                    <DetailsRating />
+                    {isLoggedIn && (
+                      <NavLink
+                        to={profile ? `/give-advice/?id=${id}` : '#'}
+                        onClick={handleNotConnected}
+                      >
+                        <button
+                          className="btn btn-primary btn-primary:hover"
+                          type="button"
+                        >
+                          Je donne mon avis
+                        </button>
+                      </NavLink>
+                    )}
+                    {!isLoggedIn && <ModalSignIn />}
 
+                    <div className="flex mt-3 justify-center">
+                      <div className="mr-3">
+                        <TwitterShareButton
+                          url={shareUrl}
+                          title={title}
+                          className="Demo__some-network__share-button"
+                        >
+                          <TwitterIcon size={32} round />
+                        </TwitterShareButton>
+                      </div>
+                      <div>
+                        <FacebookShareButton
+                          url={shareUrl}
+                          title={title}
+                          className="Demo__some-network__share-button"
+                        >
+                          <FacebookIcon size={32} round />
+                          <div>
+                            <FacebookShareCount
+                              url={shareUrl}
+                              className="Demo__some-network__share-count"
+                            >
+                              {(count) => count}
+                            </FacebookShareCount>
+                          </div>
+                        </FacebookShareButton>
+                      </div>
+                    </div>
+                  </div>
+                </div>
+              </div>
               <div className="container">
                 <p className="h-auto border border-grey p-5">
                   <span className="text-xl">Composition : </span>
@@ -195,7 +277,9 @@ export default function ProductInfo() {
                       Ratio calcium/phosphore :{' '}
                       {foodDetails.phosphorus_100g &&
                       foodDetails.calcium_100g ? (
-                        foodDetails.calcium_100g / foodDetails.phosphorus_100g
+                        (
+                          foodDetails.calcium_100g / foodDetails.phosphorus_100g
+                        ).toFixed(2)
                       ) : (
                         <span className="italic text-xs"> Inconnu</span>
                       )}{' '}
